parsing 위치바꾸기
