+++ Client/myListOpen.c
@@ -1,7 +1,6 @@
 #include "myListOpen.h"
-
-int myListOpen(){
-//int main(void){ //myListOpen();
-	system("/bin/ls");
+ 
+int main(void){ //myListOpen();
+	system("/bin/ls -alR > myList.txt");
 	return 0;
 }
