--- Server/myList.txt
@@ -1,44 +1,0 @@
-.:
-total 128
-drwxrwxr-x 4 linux linux  4096 Jul  7 22:55 .
-drwxrwxr-x 8 linux linux  4096 Jul  7 19:49 ..
--rwxrwxr-x 1 linux linux  8304 Jul  7 22:55 a
--rw-rw-r-- 1 linux linux    55 Jul  7 19:18 account.txt
--rwxrwxr-x 1 linux linux 22816 Jul  7 19:49 fileDanzi
--rw-rw-r-- 1 linux linux  4727 Jul  7 19:55 fileDanzi.c
--rw-rw-r-- 1 linux linux   455 Jul  7 19:18 fileDanzi.h
--rw-rw-r-- 1 linux linux  1422 Jul  7 19:18 fileUpload.c
--rw-rw-r-- 1 linux linux   144 Jul  7 19:18 fileUpload.h
--rw-rw-r-- 1 linux linux   836 Jul  7 19:55 getRequest.c
--rw-rw-r-- 1 linux linux   268 Jul  7 19:18 getRequest.h
-drwxrwxr-x 2 linux linux  4096 Jul  7 19:55 home
--rw-rw-r-- 1 linux linux  2644 Jul  7 19:18 listOpen.c
--rw-rw-r-- 1 linux linux   200 Jul  7 19:18 listOpen.h
--rwxrw-rw- 1 linux linux   218 Jul  7 22:53 list.txt
--rw-rw-r-- 1 linux linux   108 Jul  7 22:55 myListOpen.c
--rw-rw-r-- 1 linux linux   105 Jul  7 19:18 myListOpen.h
--rw-rw-r-- 1 linux linux     0 Jul  7 22:55 myList.txt
-drwxrwxr-x 2 linux linux  4096 Jul  7 19:18 ncursesLib
--rw-rw-r-- 1 linux linux     0 Jul  7 19:18 request.txt
--rw-rw-r-- 1 linux linux  2559 Jul  7 22:53 response.c
--rw-rw-r-- 1 linux linux   263 Jul  7 19:18 response.h
--rw-rw-r-- 1 linux linux   146 Jul  7 19:18 serverMain.c
--rw-rw-r-- 1 linux linux   109 Jul  7 19:18 serverMain.h
--rw-rw-r-- 1 linux linux  3534 Jul  7 20:15 serverStart.c
--rw-rw-r-- 1 linux linux   675 Jul  7 19:18 serverStart.h
-
-./home:
-total 20
-drwxrwxr-x 2 linux linux 4096 Jul  7 19:55 .
-drwxrwxr-x 4 linux linux 4096 Jul  7 22:55 ..
--rw-rw-r-- 1 linux linux   39 Jul  7 19:18 fileDanzi2.txt
--rw-rw-r-- 1 linux linux   28 Jul  7 19:18 fileDanzi.txt
--rw-rw-r-- 1 linux linux  270 Jul  7 22:08 list.txt
-
-./ncursesLib:
-total 44
-drwxrwxr-x 2 linux linux  4096 Jul  7 19:18 .
-drwxrwxr-x 4 linux linux  4096 Jul  7 22:55 ..
--rw-rw-r-- 1 linux linux  1592 Jul  7 19:18 clientNcurses.c
--rw-rw-r-- 1 linux linux 28622 Jul  7 19:18 commonNucurses.c
--rw-rw-r-- 1 linux linux  2746 Jul  7 19:18 commonNucurses.h
