+++ Server/myListOpen.c
@@ -1,6 +1,6 @@
 #include "myListOpen.h"
  
 int main(void){ //myListOpen();
-	system("/bin/ls -alR > myList.txt");
+	system("/bin/ls");
 	return 0;
 }
